Stop converting from UTF-8 to Windows encoding.
This conversion, I think, was necessary for Anki 1 importing but
Anki 2 seems to import UTF-8.

This change means default display of the output in Excel is broken,
but the underlying CSV file is good UTF-8.

The change also adds an attribute and hidden input to the form to
enforce UTF-8, based on Rails 3's form helper.